Use underline for MatchParen in some light variants.
In order to avoid reverting dc6c1df completely, this commit introduces
underline as a further clue to distinguish matching parentheses. In the
normal and high contrast variants, underline is used for MatchParen when
g:solarized_visibility is set to "low" or "normal", while a dark
background is used when g:solarized_visibility is set to "high". For
flat and low contrast variants, underline is used when
g:solarized_visibility is set to "high" (otherwise, the appearance is
the same as before this commit).