Allow the user to choose old style's Cursor highlight group.
Add a new option to choose the original Solarized's cursor style.

So, by default now the cursor is white on blue in dark themes and white
on orange/red (depending on visibility) in light themes no matter which
variant is used.

If you prefer a grey cursor, as in the original Solarized, put this in
your `vimrc`:

    let g:solarized_old_cursor_style = 1